Use Renderers factory
